Create choropleth using realistic data
It is imagined that data to be used to plot such maps would be in data tables that match values to specific states. So the function now receives a data frame and some reference to the columns of interest, as well as guide to how the values are to be categorized. This will need to be fine-tunes, of course, but it now works.

One thing that comes to mind is the best approach. Should one use the formula, which is easier to the user for capturing the variables, but a bit more involved for programming? Or should one apply the tidyverse approach for similar reasons?