Bump version 0.0.7.9010
Changes:
A new function was added to ensure the proper assigning of colours to the various polygons hat are duplicated. Prior to this fix, the colours were recycled by maps::map as the names of the regions exceeded the actual states provided.

Another function has also been defined for intelligent search for data frame colums containing states' names. It is yet to be implemented.