Remove parameter for states in choropleth mode
The states are already supplies in the first argument. Thus, the function has been given an additional routine wherein the data supplied is scanned for the existence of a column that has those states supplied in the first argument (or all states in the default argument).

Tests were adjusted appropirately.