Enable selection of specific duplicated polygons
Essentially what happeded was that before this change, all the duplicated polygons' names (i.e. labels) were replaced with empty strings, barring the first in the list. Unfortunately, this led to a situation where the selected polygons fell into a zone of the map that close to, or even overlapping, with neighbouring regions. Some of the labels were obscured. So some hard-coding was introduced to set the label in a more central position. This is a temporary fix.

So other functions were introduced to get this done, as well as unit tests.

rlang package is written onto the NAMESPACE and magrittr is also being used for more compact code. This is something that will be used more pervasively in the package over time.