Enable deductive choropleth mapping
The drawing of choropleth maps is no longer initiated by explicit arguments, but deduced from the availability of key data such as a data frame and at a minimum, a vector bearing the values to be displayed or a reference to a column that contains such values.

Plans for fine-tuning, include ability to differentiate between factors and numerical data, default values for breaks and/or bins, and refinements in the use of legends.