Bump version to 0.1.0
Principally, the most important change in the package is the ability to
plot more compelling and data-oriented maps such as GIS coordinates
(points) and choropleths (polygons). The internal workings of the package
have received a significant increase in complexity and hundreds of unit
tests were added. Also, attempts were made to prepare the ground for
future changes that are already planned.