Remove error check for non-character vectors
In the function 'is_state' we get an error when we check a vector that is not even a character vector. This may not be too useful for functional programming, where we may want to traverse a vector/data frame looking for States. So this has been reviewed: Instead, the function has an early return evaluating to 'FALSE' when the input is found not to be a character vector.

Approprate unit tests were included.