Change `is_state` API
The `is_state` function has been fundamentally modified, such that it is simplified to just return a logical vector for element-wise test of an object for the States. All the other arguments for selecting elements, checking NAs have been deprecated in this version. Those other actiions are to be done using other base functions such as `all`, `anyNA`.

The function throws a warning, when a missing value is among the elements. It works only for atomic vectors, throwing an error when this is not the case or when `NULL` is passed to it.

Documentation and unit tests were equally updated.