Refactor state functions
Version 0.1.3.9007

Essentially, what I did was to kind of abstract away the internal object that is created for the states and put it into a function that creates it as a named list or optionally as a character vector.

For `fix_state()` input validation was changed slightly to ensure that only objects of type character are accepted (and not *type* which would allow things like matrices or arrays to be accepted). Of course, there is a slim possiblity that this would be reversed in the future, if there is any merit in testing matrices - because in the event that we do have a character matrix of States, there would be no way of testing it without unclassing it or dropping the `dim` attribute. Relevamt unit tests have been added to this effect.

The documentation has been modified with a new entry for the `fix_state()` function.