Return non-charactoer object on testing for States
The thing is, if we get to pass an object that is not a character vector,
there really is no point testing whether it's a States. Furthermore,
atomic vectors are all of the same time, so it's doesn't make any sense to
return an all-FALSE logical vector when such an object is tested. So we
just skip it entirely, with a warning. On the other hand, when we have a
character vector longer than 1, each element is checked whether it is a
State, and returning a logical vector accordingly.`:wq