Fix broken unit tests
There was an issue that came up when, in the maps::map... family of functions, the `plot` argument evaluates to FALSE. Since we want to retain the ability to return non-plotted maps objects, it is necessary to carefully fix this and retain the behaviour. Essentially, we applied are relatively dirty fix, precursor to planned refactoring with an OOP based approach to handling the complexity of the mapping features of the package.