Improve label collection
When trying to map LGAs, and because they are so many, the function, in trying to select the name/label of the LGA to be mapped, does some partial matching. Thus, when trying to draw the map of, say, 'Abak' in Akwa Ibom State, the function tries to also draw the map of 'Abakaliki' in Ebonyi state.

The first change made is to provide anchors for the pattern matching when trying to extract the names/labels from the spatial data object. Secondly, in the mapping function itself, we have made the argument `exact` to be `TRUE`, thus requiring an exact match for any name passed.

This challenge exposed a soft underbelly in the package: The fact that shapefiles may not contain accurate metadata linked to the maps being drawn. This will have to be addressed. A possible fix is to have a function that runs a check for any shapefile data introduced; however, before this can be done, we have to be 100% sure of the authority of the names we are using as a baseline.