Enable repair of shapefile data
The shapefiles I have are riddled with errors, which now provides an
opportunity to develop constructs for repairing them as well as making
some other relatively low-level changes in the data structures. The focus
was mostly on the LGAs. Repair of mispelt names is now enhanced and works
pretty well at the moment, doing both exact and partial matching.
Informative messages are added to guide the user who is repairing LGA
names.

Internal data are now generated from the shapefile using classes and
methods from the 'sp' package. We're almost there...