Enable interactive repair of LGA names
Since the LGAs are so many, and to make it less cumbersome, when the automatic repair cannot be done by `fix_region` and interactive option has been included. The user is guided step-by-step by providing options for repair, and when this is not feasible, the user can skip the repair process.

This has been used to update the internal LGA data in the shapefile.