Reorder the logic
Actually what I wrote before made absolutely no sense - allowing the function to go through all the motions only to fail in some instances on the last statement. Cringe. Fixed that now.

The failure of a unit test was repaired, which occured due to the pretty quotation marks used in one of the condition messages, that wasn't initially reproduced in the test. Though it is baffling that this test is failing for the first time. Did the base function `SQuote()` change recently?