Develop unit tests
There's something seriously wrong with the way LGAs are extracted and this is really mucking up the maps. Working on creating unit tests first - this time around we are going to test each and every State and LGA. Focus will first be on the numbers. Then we will look at how the mapping function extracts it. Later we could set up something for the coordinates.