Fix LGA mapping problem
The initial thinking that the with plotting the map for Borno State was due to an issue with the coordinates. It later turned out that the problem was that there was partial mapping of region (i.e. LGA) names and this led to the mapping of regions that actually belonged to other States. The fix that was eventually applied involved the use of regular expressions partially match the regions but to also tighten the accuracy of the pattern matching such that only actual LGA names would be checked. To be specific, this meant that for an LGA like "Kaga" which was matching "Kagarko" this would not work any longer - but where there were multi-polygons, "Kaga:1" and "Kaga:2" would match.