Add features to lgas constructor
Because of the those LGAs that bear the same name with their State, a new argument `strict` was added. When set to `TRUE`, an individual region name that has a State synonym will be set as an lgas object. Otherwise, the object will contain all the LGAs in that State.

Also at `State` attribute is added to the object created. This is to enable the idenfication of which State LGAs are coming from. This is especially helpful when drawing State LGA maps; the attribute is used to subset the relevant spatial data to prevent name clashes (there are LGAs whose names are synonymous across various States).