Reset heading in single LGA console output
When the single LGA objects are printed out, notably where we force-construct synonymous LGAs, the printout was producing a header for 'States' instead of 'LGAs'. The logic in the the print.regions method was modified slightly to overcome this issue.

Also some additional unit tests were added to ensure the proper definition of single State/LGA object for those that are synonymous.