Fix bug in reporting on LGA spelling repairs
It was noticed that when the spelling error involved an issue with the letter casing, no reports where printed. The bug was found to be in line 152 of the file 'methods-fix_region.R', where in matching the available names, the argument `ignore.case` was set to `TRUE`. This has been undone and the function now behaves as expected.

Also a note was added to the documentation to inform the use that when only a single string was supplied misspelt, automatic fixed are not available for the function and thus an error is signalled instead. Example code and unit tests were equally added.