Update interactive LGA spelling fixing
Generally, the change was focused on improving on the output and condition signalling. Internally, the attributes used to convey the changes made in the LGAs while corrections are made was tweaked for the interactive operations. The significance of this is in the generation of reliable outputs on what is going on per time.