Update repair of misspelt regions
Essentially, the approach to fixing spelling mistakes was reviewed. Provision of entirely mispelt character vectors is an error condition that can be fine-tuned by using `states()` or `lgas()` function, depending on that type of region is being repaired. When there is a mix of corretly and mispelt regions, the class can be more easily be deduced for automatic repairs. Where there are more than one possible match for any given region name, this will be left undone and the user has the option of doing the repair interactively by following prompts and using suggested new values.