Update unit tests with new function
A new function has been added that will be exported in the next release. It's job is to deal with situations where an LGA is found in more than one State by disambiguating it. This is important for drawing LGA level maps. For now the function is used to ensure those tests pass.

Also some positional switching of parameters for the `lgas` function was dome to make it a little more intuitive.