Fix poor output behaviour
When printing out messages on LGAs whose spellings have been corrected, repetitions of the same old-new word combination were printed to the console, which is totally unnecessary and very unwieldy when there are so many changes made. The adjustments that were made principally involve selection of only the unique names of LGA inputs corrected.