Fix bug in fix_mobile function
When supplied only NAs, this function fails and this is particularly
problematic when one is iterating over a data frame in which the column
being checked might just happen to be empty.

After applying the fix, the function works well with missing values and
returns the argument as-is in that situation. Additional tests were
added to confirm the assertions.