Include infrastructure for checking shapefile data
There are issues with the spelling of LGAs and the problem is
deep-seated. In fact, it has been noted that there a clear mistakes in
the spellings wihin the main source document, which is the 1999
Constitution. So, our approach to dealing with this errors has to be
reviewed/overhauled to provide a measure of consistency.

Internal functions have been added that will make it easier to scan
shapefiles for LGA mismatches and subsequently dealing with them. This
will be made applicable to whatever shapefiles are used in the future,
and the approach will be extended for this purpose.