Draw down data on LGAs from Wikidata
I have decided to use the data from this source for the LGAs, since it
has become apparent that there are mistakes in the Constitution
document, as well as other discrepancies in State-level data posted on
websites. Worse, only few States have the data posted online. The
Wikidata has references and it's something that can be improved, if
needed. Also, the data is structured, so the script being used will have
a degree of longevity.

The data that I have gotten so far will need a little cleaning, but I
will start that process by editing on Wikidata. I am considering
maintaining a separate database, if need be.