Implement improvements to package and internal data
A lot more care has been put into getting correct and up to date data on LGAs from external sources as well as making necessary adjustments to the same LGA data elements within the spatial data used by the package's functions. 2 scripts in the `data-raw` directory perform these tasks. Several utility functions were written afresh or updated to ease the process and to future-proof the workflow. This is the first iteration - other enhancements will be included. Note that there is currently a bill being considered in the legislature that renames several of these LGAs, so there's more work ahead.