Update DESCRIPTION Imports entry for `{sf}`
`R CMD check` failed for Ubuntu due the inability to resolve the dependency on the `{sf}` package. On closer examination,  I noticed that the package version on CRAN is dash-separated (x.x-x), while in the DESCRIPTION file is is dot-separated. This change is being made to see whether it will fix the problem with `check`.