Undo spelling fix on attribute
I had decided to change the spelling of 'misspelt' to American English 'misspelled'. This I did using 'Replace All' and the attribute so named was also affected. This messed with the unit tests and, if published, could potentially break some users' code. So, I have maintained that spelling for the attribute while the documentation will use the Americanized version.