Fix failing internal data access
The S3 constructor `lgas()` was failig whenever it was called with the
`::` operator when the package namespace is not already loaded. This
failure occured because the function uses the `lgas_nigeria` dataset,
which when the package is not attached with `library`, is only loaded
within the function environment. Unfortately there are a few
non-exported helper functions that are called when we attempt to extract
LGA data for selected States (those functions were not seeing the
internally loaded data. So, I decided to do away with the functions in
order to expose the data to the logic. In fact, the functions were quite
unnecessary and it helped to reveal some of the undue complexity in the
packages code--something that will be addressed later.

Other changes made include:

- Improved sorting of the output of the `lgas()` function.
- Code refactoring
- Additional unit tests